fixes genai build.dart script
